requirements.txt
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
openai==1.14.1
python-dotenv==1.0.1
SQLAlchemy==2.0.28
streamlit==1.32.2